STEP 10. Break up event handlers.
