+++ altitude/static/js/frontend-app.js
@@ -12,6 +12,7 @@ import {
 } from "./common/dragon-drop.js"
 import { showAssetDetailModal } from "./common/modal.js"
 import { hydrateAppFragments } from "./fragments/index.js"
+import { registerAppEventListeners } from "./listeners/index.js"
 import { setImgSrcAndWait } from "./search-results/detail-navigator.js"
 import "./search-results/dragon-drop.js"
 
@@ -148,285 +149,7 @@ export class FrontendApp {
     }
 
     registerEventListeners() {
-        document.body.addEventListener(Const.events.folderMoved, (event) => {
-            const movedFolderId = event.detail["movedFolderId"]
-            const newParentId = event.detail["newParentId"]
-
-            if (newParentId === movedFolderId) {
-                return
-            }
-
-            const movedFolder = new Folder(movedFolderId)
-            const newParent = new Folder(newParentId)
-            const oldParent = movedFolder.parent()
-
-            htmx.ajax(
-                "put",
-                `/htmx/folder/r/${this.context.getRepoId()}/move?movedFolderId=${movedFolderId}&newParentId=${newParentId}`,
-                {
-                    swap: "none",
-                    handler: (response) => {
-                        const status = response["htmx-internal-data"].xhr.status
-
-                        if (status === 200) {
-                            movedFolder.closeContextMenu()
-                            movedFolder.clearChildren()
-
-                            newParent.incrementNumOfChildren()
-                            oldParent.decrementNumOfChildren()
-
-                            showSuccessSnackBar(
-                                `Folder ${movedFolder.name()} moved into "${newParent.name()}"`,
-                            )
-
-                            if (newParent.isExpanded()) {
-                                newParent.addChild(movedFolder)
-                                movedFolder.collapse()
-                            } else {
-                                movedFolder.remove()
-                            }
-
-                            newParent.updateVisualState()
-                            oldParent.updateVisualState()
-                        } else if (status === 409) {
-                            showWarningSnackBar(
-                                response["htmx-internal-data"].xhr.responseText,
-                            )
-                        } else {
-                            showErrorSnackBar(
-                                `Error moving folder "${movedFolder.name()}". Status: ${status}`,
-                            )
-                        }
-                    },
-                },
-            )
-        })
-
-        document.body.addEventListener(Const.events.folderAdded, (event) => {
-            const parentFolder = new Folder(event.detail["parentId"])
-            parentFolder.incrementNumOfChildren()
-            parentFolder.updateVisualState()
-        })
-
-        document.body.addEventListener(Const.events.folderDeleted, (event) => {
-            const folder = new Folder(event.detail["id"])
-            const parent = folder.parent()
-
-            parent.decrementNumOfChildren()
-            parent.updateVisualState()
-
-            showSuccessSnackBar(`Folder "${folder.name()}" deleted`)
-            folder.remove()
-        })
-
-        document.body.addEventListener(
-            Const.events.confirmPersonMerge,
-            (event) => {
-                const mergeSourceId = event.detail["mergeSourceId"]
-                const mergeDestId = event.detail["mergeDestId"]
-
-                if (mergeSourceId === mergeDestId) {
-                    return
-                }
-
-                htmx.ajax(
-                    "GET",
-                    `/htmx/people/r/${this.context.getRepoId()}/modals/merge`,
-                    {
-                        swap: "innerHTML",
-                        target: "#modalContent",
-                        values: { ...event.detail },
-                    },
-                )
-            },
-        )
-
-        document.body.addEventListener(Const.events.personMerged, (event) => {
-            const mergeSourceId = event.detail["mergeSourceId"]
-            const sourcePersonEl = htmx.find(`#person-${mergeSourceId}`)
-
-            if (sourcePersonEl) {
-                sourcePersonEl.remove()
-            }
-
-            showSuccessSnackBar("Person merged successfully")
-        })
-
-        document.body.addEventListener(
-            Const.events.personNameEdited,
-            (event) => {
-                const personId = event.detail["personId"]
-                const newPersonName = event.detail["newPersonName"]
-                const personNameEl = htmx.find(`#person-${personId} .name a`)
-
-                if (!personNameEl) {
-                    return
-                }
-
-                personNameEl.textContent = newPersonName
-                personNameEl.classList.remove("unknown")
-            },
-        )
-
-        document.body.addEventListener(
-            Const.events.personCoverFaceSet,
-            (event) => {
-                const personId = event.detail["personId"]
-                const faceId = event.detail["faceId"]
-                const imageEl = htmx.find(`#person-${personId} .image img`)
-
-                if (!imageEl) {
-                    return
-                }
-
-                imageEl.src = `/content/r/${this.context.getRepoId()}/face/${faceId}`
-            },
-        )
-
-        document.body.addEventListener(
-            Const.events.personMarkedAsBadMatch,
-            (event) => {
-                const personId = event.detail["personId"]
-                const personEl = htmx.find(`#person-${personId}`)
-
-                if (personEl) {
-                    personEl.remove()
-                }
-            },
-        )
-
-        document.body.addEventListener(Const.events.assetMoved, (event) => {
-            const assetId = event.detail["assetId"]
-            const folderId = event.detail["folderId"]
-            const selectedAssetsStore = this.Alpine.store(
-                Const.state.selectedAssets,
-            )
-
-            if (
-                !selectedAssetsStore.isEmpty &&
-                selectedAssetsStore.contains(assetId)
-            ) {
-                this.dispatch(Const.events.batchAssetsMoved, { folderId })
-                return
-            }
-
-            this.moveAssets({ folderId, assetIds: [assetId] })
-        })
-
-        document.body.addEventListener(
-            Const.events.batchAssetsMoved,
-            (event) => {
-                const folderId = event.detail["folderId"]
-                const selectedAssetsStore = this.Alpine.store(
-                    Const.state.selectedAssets,
-                )
-
-                this.moveAssets({
-                    folderId,
-                    assetIds: Array.from(selectedAssetsStore.items.keys()),
-                })
-            },
-        )
-
-        document.body.addEventListener(Const.events.assetTrashed, (event) => {
-            const assetId = event.detail["assetId"]
-            const selectedAssetsStore = this.Alpine.store(
-                Const.state.selectedAssets,
-            )
-
-            if (
-                !selectedAssetsStore.isEmpty &&
-                selectedAssetsStore.contains(assetId)
-            ) {
-                this.dispatch(Const.events.batchAssetsRecycled)
-                return
-            }
-
-            this.recycleAssets({ assetIds: [assetId] })
-        })
-
-        document.body.addEventListener(
-            Const.events.batchAssetsRecycled,
-            () => {
-                const selectedAssetsStore = this.Alpine.store(
-                    Const.state.selectedAssets,
-                )
-
-                this.recycleAssets({
-                    assetIds: Array.from(selectedAssetsStore.items.keys()),
-                })
-            },
-        )
-
-        document.body.addEventListener(Const.events.batchAssetsPurged, () => {
-            const selectedAssetsStore = this.Alpine.store(
-                Const.state.selectedAssets,
-            )
-
-            this.purgeAssets({
-                assetIds: Array.from(selectedAssetsStore.items.keys()),
-            })
-        })
-
-        document.body.addEventListener(Const.events.batchAssetsRestored, () => {
-            const selectedAssetsStore = this.Alpine.store(
-                Const.state.selectedAssets,
-            )
-
-            this.restoreAssets({
-                assetIds: Array.from(selectedAssetsStore.items.keys()),
-            })
-        })
-
-        document.body.addEventListener(
-            Const.events.viewSettingChanged,
-            (event) => {
-                this.handleViewSettingChanged(event)
-            },
-        )
-
-        document.body.addEventListener(Const.events.deselectAll, () => {
-            const selectedAssetsStore = this.Alpine.store(
-                Const.state.selectedAssets,
-            )
-            selectedAssetsStore.items.forEach((asset) => {
-                asset.deselect()
-            })
-            selectedAssetsStore.items.clear()
-        })
-
-        document.body.addEventListener(Const.events.showNext, () => {
-            this.handleShowNext()
-        })
-
-        document.body.addEventListener(Const.events.showPrevious, () => {
-            this.handleShowPrevious()
-        })
-
-        document.body.addEventListener(Const.events.detailShown, (event) => {
-            this.Alpine.store(Const.state.shadowResults).currentAssetId =
-                event.detail.assetId
-        })
-
-        document.body.addEventListener(Const.events.escapeKeyPressed, () => {
-            this.handleEscapeKeyPressed()
-        })
-
-        document.body.addEventListener("htmx:afterRequest", (event) => {
-            this.handleAfterRequest(event)
-        })
-
-        document.body.addEventListener("htmx:afterSwap", (event) => {
-            this.handleAfterSwap(event)
-        })
-
-        document.body.addEventListener("htmx:beforeRequest", (event) => {
-            this.handleBeforeRequest(event)
-        })
-
-        document.body.addEventListener("htmx:load", (event) => {
-            this.handleHtmxLoad(event)
-        })
+        registerAppEventListeners(this)
     }
 
     handleBeforeRequest(event) {
