+++ altitude/static/js/listeners/index.js
@@ -1,0 +1,12 @@
+import { registerFolderListeners } from "./folders.js"
+import { registerPeopleListeners } from "./people.js"
+import { registerAssetListeners } from "./assets.js"
+import { registerHtmxAndSearchListeners } from "./htmx-search.js"
+
+export function registerAppEventListeners(app) {
+    registerFolderListeners(app)
+    registerPeopleListeners(app)
+    registerAssetListeners(app)
+    registerHtmxAndSearchListeners(app)
+}
+
